[TASK] Remove extension class for ElementBrowserFolderTreeView
EXT:recordlist ships with an extended ElementBrowserFolderTreeView
which is a subclass of the regular ElementBrowserFolderTreeView from
EXT:backend. The subclass can be easily merged into the parent class
as the logic is non-breaking for the areas which use the parent class.

The patch removes the child class (4th level inheritance!) into the parent
class and changes all occurences in the core to the parent class.

A class alias map is added for the old class into EXT:compatibility6.

Resolves: #67999
Releases: master
Change-Id: Id13f999311230a8aafbc2f61f4538c5e55d052c6
Reviewed-on: http://review.typo3.org/41024
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>