[FEATURE] Use Symfony Console in CliRequestHandler
The logic for outputting the default checks when calling cli_dispatch.phpsh is now using Symfony Console showing
colored error messages when entering wrong cliKeys.

The global variable $GLOBALS['MCONF']['name'] which was
mis-used to transfer information between the CLI RequestHandler
and the BackendUser (only when in CLI mode) for authentication
was removed, as all the logic is now available inside the CliRequestHandler and can be resolved there already.

The call "$BE_USER->checkCLIuser()" on every backend request
is now removed and the method itself is now marked for deprecation,
as the same functionality (except for using exceptions instead of fwrite() calls) is now in CliRequestHandler.

The two constants TYPO3_cliKey and TYPO3_cliInclude have
been marked internally as deprecated, as they are not in use
anymore.

Resolves: #68804
Releases: master
Change-Id: Ia7730d43dd970fd33daeb78e538c8d4deb741d93
Reviewed-on: http://review.typo3.org/42295
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Susanne Moog <typo3@susannemoog.de>
Tested-by: Susanne Moog <typo3@susannemoog.de>