[!!!][BUGFIX] Fix wrong usage of Enum in InformationStatus Enum itself
The Enumeration implementation of InformationStatus provides
a mapStatusToInt() method that must expect a InformationStatus instance
as parameter instead of a string.

The method mapStatusToInt() has also been replaced by "isGreaterThan" as
a comparison method fits better into the enumeration usage.

This change is breaking but should not really break anything as it
was introduced in 7.4 and only used at one place in the core.

Change-Id: I84548ba291360d1b20a607fa1e5149b95ca93bda
Resolves: #69224
Releases: master
Reviewed-on: http://review.typo3.org/42804
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Tested-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Sascha Egerer <sascha@sascha-egerer.de>
Tested-by: Sascha Egerer <sascha@sascha-egerer.de>