[TASK] Reduce PageLayoutController and PageLayoutView dependencies
This is a follow-up to #68395 to reduce complexity and exposed
public API of that change.

Even if class components have the names "controller" and "view",
which implies an underlying MVC-stack, the PageLayout components
are far from using the pattern. That's why it's quite fine to put
processing logic to the place where it makes most sense and where
dependencies and cross-calls to other objects are kept low.

This change is only about refactoring and reducing the public API
in terms of information hiding and encapsulation. Besides that
the $languagesInColumnCache class variable was defined but never
use - which is changed now as well. Further flaws concerning
proper record selection are part of another change.

Related: #68395
Resolves: #69261
Releases: master
Change-Id: I8bb7737d154a5d2c7a584dbe11fd0a5abfabafba
Reviewed-on: http://review.typo3.org/42838
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>