[FEATURE] New option to disable the age display of dates per field by TCA
If an input field of a table is declared as "input" and evalued with
"date", record lists show the date with an automatically
calculated age. This is not needed, when a user only wants to see the date
and no further information about the age related to now.

This change enables integrators to disable the hard coded calculation by
a new TCA option called "disableAgeDisplay". Par example, to remove age
calculation for tt_content and the field "date", integrators can add the
following lines to the TCA:
$GLOBALS['TCA']['tt_content']['columns']['date']['config']
['disableAgeDisplay'] = true;

So, disabling the age display can be configured per table and field.

Resolves: #28243
Releases: master
Change-Id: Id2ce6512764881889b4181f73db36048d5704be8
Reviewed-on: http://review.typo3.org/42849
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Mats Svensson <mats@pixelant.se>
Tested-by: Mats Svensson <mats@pixelant.se>
Reviewed-by: Philipp Gampe <philipp.gampe@typo3.org>
Tested-by: Philipp Gampe <philipp.gampe@typo3.org>