[TASK] Cleanup navigation frame module registration
http://review.typo3.org/37611 made the navigation frame dispatched.
For this it leveraged the addNavigationComponent API,
which was introduced only for ExtJS tree components.

Since we want to get rid of ExtJS (and this API specifically),
we need to introduce a different way to make modules or routes
navigation components.

This change reverts the above changes to the API and instead uses
the `navigationFrameModule` option which was introduced in
https://review.typo3.org/#/c/30593

To make this option work with top level modules and inheritance of
the navigation frame,
the evaluation of the option is moved to the module loader, which
has the benefit to have the code only in one place while it needed
to be in two places before.

Resolves: #69270
Releases: master
Change-Id: I7cb2ca2ee6d04001af424c88a1db828d1712caa1
Reviewed-on: http://review.typo3.org/42856
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>