[TASK] FormEngine: Remove name="foo[bar]_hr" from input fields
Update the handling of human readable fields so that the values are not
being set and only the hidden field with the actual value is submitted.

This patch also updates the handling for the _list and _mul appendixes
as these use the same methods and needed to be changed as well.

Resolves: #69672
Releases: master
Change-Id: I7a92c924a567621c2f79c736f878af288595c657
Reviewed-on: http://review.typo3.org/43254
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>