[TASK] Improve session timeout dialog
This renames the "Abort" action to "Stay logged in" and makes it
the default in the session timeout dialog. Thus one can simply
press the Space/Return key to stay logged in which is the most
likely action in this situation.

This drops the translation label "mess.refresh_login_abort_button"
from locallang_core.xlf which is now unused. Also the label
"mess.refresh_login_confirm_button" is changed to
"mess.refresh_login_logout_button".

Resolves: #69774
Releases: master
Change-Id: I71b797fb61750a293718e144bc76e28199874c38
Reviewed-on: http://review.typo3.org/43288
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Michael Oehlhof <typo3@oehlhof.de>
Tested-by: Michael Oehlhof <typo3@oehlhof.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>