[TASK] Have eIDs with PSR-7 without ControllerInterface
In order to allow the same logic as with the Routing and the
direct information which method to call, the ControllerInterface
is removed again.

Additionally, the previously introduced Dispatcher is now renamed
to RouteDispatcher (based on routing), and a simple dispatcher
is now added to EXT:core, which the RouteDispatcher derives from.

Resolves: #69846
Releases: master
Change-Id: Ica028f81aa377fd60e5159907c8c226fe0ebd34a
Reviewed-on: http://review.typo3.org/43352
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>