[BUGFIX] Show "add pages" button to qualified users only
If an editor has no rights to create new pages of any type the
icon for showing the document toolbar for the drag'n drop
adding is shown nevertheless.

This patch changes this behavior. The icon is only shown
after the AJAX call returns at least one allowed doktype.

Change-Id: I35050e4cfe923b58c26d85ce39a69dc69aa58515
Resolves: #26901
Releases: master
Reviewed-on: http://review.typo3.org/43629
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>