[BUGFIX] Evaluate displayConditions based on array values
With the rewrite of the FormEngine, all data defined as select
fields in TCA became arrays in the data provision. Therefore
displayConditions based on single values are not evaluated
correctly anymore.

A check was added, when the array contains only one value, this
one is used to perform the evaluation.

Change-Id: If43067eb5d19748afce02b41ce44a14ca4b7fa1f
Resolves: #70265
Releases: master
Reviewed-on: http://review.typo3.org/43693
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>