[BUGFIX] Revert BackendActionController in favor of BackendView
The module template is a view component and should not clutter
the controller code of a module or at least only in a minimal way.

Now we introduce a BackendView which decorates a fluid template view.
This can be easily used in backend modules by using the defaultViewObjectName API

Releases: master
Resolves: #70358
Change-Id: If164c2e6774bd75e414c6417e0d8a4e989fe08e6
Reviewed-on: http://review.typo3.org/43797
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Reviewed-by: Michael Oehlhof <typo3@oehlhof.de>
Tested-by: Michael Oehlhof <typo3@oehlhof.de>
Reviewed-by: Benjamin Kott <info@bk2k.info>
Tested-by: Benjamin Kott <info@bk2k.info>