[BUGFIX] Restrict parent calculation for inline ajax
If a child is expanded, added or synchronized in inline ajax call,
mostly only the inline tca configuration of the parent record is
needed, but not the whole thing.

The patch adds a dedicated InlineParentRecord data group that only
calls needed data providers to achieve this. This has the huge benefit
that the record is no longer fetched from db and only one single
field is prepared. This improves performance of these ajax calls.

Additionally, the patch solves an exception adding new image records
to a new tt_content record without persisting the tt_content record
before.

Resolves: #70549
Releases: master
Change-Id: I1312d00b77446eb89f5da3b898cdff975f057eae
Reviewed-on: http://review.typo3.org/43941
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>