[TASK] FormEngine result array cleanup for internal data
Some of the data created within the FormDataCompiler stuff by
the data providers is used only within the data creation
process internally and should not be used by the render
engine later.
An example is the "inline overrule" array that is merged
by one provider to the target TCA and render engine must
not access or handle this later on.
The patch introduces a list of keys in FormDataCompiler
that is unset from the result array before it is returned
to the calling controller.

Change-Id: Ic6abf7787d18a92753ae1415e91f50411ca149d4
Resolves: #70552
Releases: master
Reviewed-on: http://review.typo3.org/43944
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>