[BUGFIX] FormEngine: Inline new intermediate record placeholders
The patch adds the inline configuration to the data compiler array
when children data is compiled instead of setting this only for
the render engine. Data provider can now use this data. The data
provider that initializes new rows is refactored and now sets
the uid of a "child-child" if an intermediate child is compiled.
This fixes a bug in the placeholder handling that now resolves
placeholder data for new inline records created by inline ajax
controller correctly.

Resolves: #70577
Releases: master
Change-Id: I7c424e159a954824a947e1df3cfbb410e15d2b78
Reviewed-on: http://review.typo3.org/43958
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>