[BUGFIX] Remove getRecordTitle() runtime cache
This reverts the runtime cache of BackendUtility::getRecordTitle()
from commit d10070e81. This low level method is getting more and
more unused anyway and is substituted with a data provider in
FormEngine for example. The cache currently introduces issues
in edge cases and the benefit of the cache in general in rather
small.

Resolves: #70435
Related: #69749
Releases: master
Change-Id: Id9f352a2193f38fafc305001a2d20db16298b2fd
Reviewed-on: https://review.typo3.org/44241
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Jan Helke <typo3@helke.de>
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>