[BUGFIX] TCA displayCond with EXT:anExt:LOADED fails
Due to the new condition evaluation construct in FormEngine, the
EXT:anExtension:LOADED display condition is very unfortunate since
the evaluation happens at the end, and depending on the configuration
of affected field, the former data provider may fail fetching data.

This is solved by deprecating this condition since it works contrary
to the "additive" TCA system: Extensions should always add or change
other extensions TCA, but should not have a "negative" dependency
to each other like "Don't do if that extension is loaded".

Core is affected at two places in combination with
static_info_tables. This also lowers the dependency from core
to static_info_tables again, which is not a good thing in
first place.

A patch for static_info_tables adding the fields there is pending.

Resolves: #70052
Releases: master
Change-Id: Iccb7a116d1f89e55e80e389a2448759eca326605
Reviewed-on: https://review.typo3.org/44251
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Jan Helke <typo3@helke.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>