[BUGFIX] Invalid CSRF token on viewing page using page-tree context menu
Frontend previews triggered using the context menu of the page-tree are
causing an exception concerning an "invalid CSRF token". The reason for
this is that URL loaded in the context menu are delivered via ExtDirect
and thus are encoded for a JavaScript context. Ampersands are encoded
as well, using the unicode serialization in JSON ('\u0026').

This issue is solved by decoding the value again and thus resolve the
expected URL to be delivered as XHR result.

Resolves: #69021
Releases: master
Change-Id: I5836ae0f3d461f6e3a091c783b9fccd4412eaffb
Reviewed-on: https://review.typo3.org/44297
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Tested-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>