[TASK] Allow to define multiple inlineLocalizeSynchronize commands
The inlineLocalizeSynchronize command in DataHandler currently
supports these formats:
* [parent][13][inlineLocalizeSynchronize] = field,14
* [parent][13][inlineLocalizeSynchronize] = field,localize
* [parent][13][inlineLocalizeSynchronize] = field,synchronize

Since the sub-commands "localize" and "synchronize" address the
whole parent scope, the particular definition of a child-id (14
in this case) is only considered for one child-element. Localizing
different children would then result in multiple independent
DataHandler processings.

The scope of this change is to extend the API to define multiple
child-ids and define the proper language, which is currently
determined from the parent-element magically.

The current string configuration format is changed to be an array,
legacy configurations are converted to the new format:

[parent][13][inlineLocalizeSynchronize] = [
  field: name of the parent field,
  language: id of the target language,
  action: either "localize" or "synchronize",
  ids: array of child-ids to be localized [1, 2, 3]
]
Either "action" or "ids" must be defined.

Resolves: #71126
Releases: master, 6.2
Change-Id: Ic7e1293a44047bfd69017e240dd8563a1dffa423
Reviewed-on: https://review.typo3.org/44354
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>