[TASK] Add menu item to clear the clipboard
The make the clipboard more usable, this patch added another menu
item to clear the clipboard. Additionally the label and message
to delete elements is now more precise.
That duplicates the function of the red button with the X, but it
seems that this function was not clear for anyone.

Resolves: #18507
Releases: master
Change-Id: I6b06c6e6f1e45c61476c37b92fd52e1e7a9eae5d
Reviewed-on: https://review.typo3.org/44392
Reviewed-by: Michael Oehlhof <typo3@oehlhof.de>
Tested-by: Michael Oehlhof <typo3@oehlhof.de>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Tested-by: Frank Nägler <frank.naegler@typo3.org>