[BUGFIX] FormEngine: Fix keepItems, addItems and removeItems handling
With the FormEngine rewrite the evaluation order and handling of the
options changed. This patch restores the previous order, adds tests
for the addItems handling as well as for the execution order of
keepItems, addItems and removeItems.

In addition the behavior of keepItems with an empty list of items has
been restored.

Resolves: #70956
Releases: master
Change-Id: I44b3036e2ba4dd824037aa689543dc2f1c653b93
Reviewed-on: https://review.typo3.org/44570
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>