[BUGFIX] FormEngine: Select correct FlexForm DS for inline records
If an inline field is defined within a FlexForm that has multiple data
structures defined the FormAjaxInlineController always selects the
default data structure due to missing context information.

This patch transmits the required context within the AJAX request to
enable selecting the appropriate data structure for the flexform.

It also changes the method to determine the path through the flexform
data structure to work in the case of an inline element configured
within a flexform within another inline element.

Resolves: #71436
Related: #70918
Related: #71564
Related: #71655
Releases: master
Change-Id: I7ecd174b78997ad3b1d1513a15a78cdc7bac23ff
Reviewed-on: https://review.typo3.org/44914
Tested-by: Armin Ruediger Vieweg <armin@v.ieweg.de>
Tested-by: Sebastian Michaelsen <michaelsen@t3seo.de>
Tested-by: Felix Rauch <rauch@skaiamail.de>
Reviewed-by: Claus Due <claus@phpmind.net>
Tested-by: Christopher Orth <corthmail@gmail.com>
Tested-by: Torsten <info@by-torsten.com>
Tested-by: Philipp Wrann <philippwrann@gmail.com>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>