[BUGFIX] Catch security exceptions in global JS module
The storage JS module is loaded in the top window.
In case this windows has been opened by some other page
(different domain) then a security exception is thrown
by the browsers, which kills the execution.

Catch exceptions when accessing the opener to avoid
this particular edge case.

Resolves: #71857
Releases: master
Change-Id: Ifa19d1f70eae71ac46c66bb759580d82ce5b6c99
Reviewed-on: https://review.typo3.org/45093
Reviewed-by: Eckard Gehrke <eckard.gehrke@gmx.de>
Tested-by: Eckard Gehrke <eckard.gehrke@gmx.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>