[BUGFIX] Determine returnUrl when deleting pages in EditDocumentController
Determine the returnUrl based on the current module and record. If the
delete action if for a record in the table "pages" the backend needs to
perform a partial reload to avoid trying to access the deleted page. In
all other cases the predetermined returnUrl is ok.

Resolves: #72058
Resolves: #72081
Related: #71425
Releases: master
Change-Id: Id019f2b4a6f60cb850322d6667094b832c62a8e0
Reviewed-on: https://review.typo3.org/45126
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>