[BUGFIX] FormEngine: Fix hiding/unhiding IRRE records with toggle icon
Fix the javascript selector used to get the field storing the hidden
state of an IRRE record so that the toggle icon changes the state again.

Also fixes reapplying the latest hidden state after expanding a
previously closed panel. If there were unsaved changes to the hidden
state of the record these got overwritten by the results of the AJAX
call to load the record details.

Resolves: #71991
Releases: master
Change-Id: I31970abab7eed877ed13a0038245b03cb569a566
Reviewed-on: https://review.typo3.org/45345
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>