[BUGFIX] Switch-User feature uses wrong user id for logging
Issue #17643 introduced logging for a simulated backend user
(switch-user) and thus logs by using the original admin user.

This is a regression in PageLayoutController::renderQuickEdit
since the used backend user id is different to the persisted
one. Besides that, all actions are performed with the permissions
of the simulated user - thus, logging with a different user seems
to wrong. The sys_log.log_data field (array) is used to transport
the information of the original (admin) user.

This change reverts the initial change of issue #17643 and adds
the initial intention of that issue as addition to the persisted
data in sys_log.log_data. All affected components are adjusted as
well to visualize that a user has been simulated (log view, record
history and lowlevel_cleaner syslog command).

Resolves: #71580
Releases: master, 7.6
Change-Id: If12df60563afd1f0746d43e62f824b20f139df8c
Reviewed-on: https://review.typo3.org/45490
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>