[BUGFIX] FormEngine: Restore select wizard for textarea elements
Add the data-formengine-input-name attribute to FormEngine TextElements
so that the wizards can locate it. As the textareas don't differentiate
between human readable and hidden fields add a check to the tbe change
handler if the form field and the human readable field are identical and
skip triggering the change event again.

Resolves: #71679
Releases: master, 7.6
Change-Id: I69e6a4f3e9f19c68bd02a0029322d52323286194
Reviewed-on: https://review.typo3.org/46032
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>