[TASK] Move BE Login BrowserLanguage detection to Locales
The CharsetConverter contains the calculation to find the
right language based on the browsers language settings
if no backend user has been logged in (Backend Login),
and chooses the TYPO3 language based on that.

This code does not belong to the CharsetConverter but
rather to the locales object.

Resolves: #73511
Releases: master
Change-Id: I4361d0e3a842efd0c5b69837776fe0c397edd593
Reviewed-on: https://review.typo3.org/46734
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Sebastian Fischer <typo3@evoweb.de>
Reviewed-by: Thomas Schlumberger <thomas@b13.de>
Tested-by: Thomas Schlumberger <thomas@b13.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>