[BUGFIX] Fix minor flaws in wizard component
After closing the wizard the internal settings object stays an object and
is not converted to an array.

Wizard.getComponent() now generates the markup if Wizard.show()
was not called before. Also, the method now returns only the generated
markup and not the (missing) modal window.

Also, the events of the wizard component are now used and not the
bootstrap modal events.

The documentation is adjusted a bit and Localization.js is fixed.

Change-Id: I0f7efd5f50247ae70fbd14600ce204c22949703d
Resolves: #73575
Releases: master
Reviewed-on: https://review.typo3.org/46788
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Tested-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>