[TASK] Cleanup Clear Cache Menu commands
The Clear Cache Menu on top of the TYPO3 Backend contains
the URLs to call tce_db which has an additional GET
parameter called "ajaxCall" which is not evaluated anymore
and can be removed.

Further cleanup could use the ajax action to fetch additional
data and evalute that data.

Resolves: #75246
Releases: master
Change-Id: I9b5f5816c5d3ac9d72cf0b382b5646e7631d0afe
Reviewed-on: https://review.typo3.org/47377
Reviewed-by: Michael Oehlhof <typo3@oehlhof.de>
Tested-by: Michael Oehlhof <typo3@oehlhof.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>