[BUGFIX] Cursor position is changed while typing
Fields in the backend are validated while typing. One validation
is the left space trim. Spaces on the left side are removed but
also the cursor position is changed. If you are a fast typer
this could end up in a mixed up input. The cursor position should
not be changed by the validation JavaScript.

Change-Id: I93e46dff763363eed193d41b1fd71fbe043fbcd8
Resolves: #75724
Related: #75626
Releases: master
Reviewed-on: https://review.typo3.org/47737
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>