[!!!][BUGFIX] Load popup configuration in case backend submodules are opened in a new tab
It's possible to open some backend modules (e.g. RTE or
EditDocumentController) in a new tab. However, the configuration for
popups is missing in such case, causing e.g. opening the "Insert image"
wizard of RTE is not possible anymore.

Place configuration inline in TYPO3.settings to not have this problem of
missing configuration.

Resolves: #76285
Releases: master
Change-Id: I1405dd90e4e00fc709d504af2ed0a936a6374fce
Reviewed-on: https://review.typo3.org/48276
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>