[!!!][TASK] Cleanup ContextMenu classes
As a pre-patch for the cleanups in the page tree, this patch removes unused
classes and simplifies the class structure. All this classes never used
outside the core.
A search in the TER doesn't find any usages.
For this reason, we decided to not deprecate the classes but made a
breaking change.

Resolves: #76527
Releases: master
Change-Id: I5a180999d5fd72dca8467e986e5870071091533c
Reviewed-on: https://review.typo3.org/48511
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>