[BUGFIX] Check for form changes on exit
Due to event juggling, setting the dirty state of a form does not work
for the first changed element. Because of this, the form is checked for
changes if a user clicks on the close button.

The original behavior was introdced with commit 003f9903 which is
reverted with this patch.

Resolves: #76893
Related: #45348
Releases: master
Change-Id: I45c2b05e096a2e7c18b8ce0aa584fc508f52ebdb
Reviewed-on: https://review.typo3.org/48752
Tested-by: Bamboo TYPO3com <info@typo3.com>
Reviewed-by: Michael Oehlhof <typo3@oehlhof.de>
Tested-by: Michael Oehlhof <typo3@oehlhof.de>
Tested-by: Riccardo De Contardi <erredeco@gmail.com>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>