[!!!][TASK] Remove deprecated usage of 'static_lang_isocode'
Remove another usage of static_lang_isocode within FormEngine code.
This field was deprecated in v7 and came along with an update wizard,
it has been marked as breaking in an 8.0 .rst file already.
There are further usages of this field that are harder to remove,
those can be done with other dedicated patches.

Change-Id: Ic6f52124d244b418802f63d148bbc89743c7cec9
Resolves: #76992
Releases: master
Reviewed-on: https://review.typo3.org/48847
Tested-by: Bamboo TYPO3com <info@typo3.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>