[BUGFIX] Show Create Content button only for allowed languages
The Create Content button inside the Page Module is now only shown for
languages the backend-user is actually allowed to edit. That way, users
that are restricted to a specific language and not allowed to edit the
default language, do not see this button inside the default language column
any more. (Note: this patch only hides the button itself, the later processing
was already prevented before.)

Resolves: #77753
Releases: master, 7.6
Change-Id: I44ebacaf1d1d0635b0af1e7e52689f077f6fa2b4
Reviewed-on: https://review.typo3.org/49800
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Jigal van Hemert <jigal.van.hemert@typo3.org>
Tested-by: Jigal van Hemert <jigal.van.hemert@typo3.org>
Reviewed-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Tested-by: Tymoteusz Motylewski <t.motylewski@gmail.com>